add guest button to login page with redirect to home
